+++ createElem.js
@@ -1,0 +1,8 @@
+export const createElement = (tag, className) => {
+  const $tag = document.createElement(tag)
+  if (className) {
+    $tag.classList.add(className)
+  }
+
+  return $tag
+}
