+++ hitAtt.js
@@ -1,0 +1,7 @@
+export const HIT = {
+  head: 30,
+  body: 20,
+  foot: 15,
+}
+
+export const ATTACK = ['head', 'body', 'foot']
