+++ players.js
@@ -1,0 +1,25 @@
+import { attack, elHP, changeHp, renderHP } from './playerHp.js'
+
+export const player1 = {
+  player: 1,
+  name: 'Subzero',
+  hp: 100,
+  img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
+  weapon: ['snowball', 'axe', 'knife'],
+  attack,
+  changeHp,
+  elHP,
+  renderHP,
+}
+
+export const player2 = {
+  player: 2,
+  name: 'Scorpion',
+  hp: 100,
+  img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif',
+  weapon: ['chain', 'axe', 'knife'],
+  attack,
+  changeHp,
+  elHP,
+  renderHP,
+}
